refactor/rename updates and DishCard.kt Logic Build - From eatas to localflav name. - Added preview to DishCard
