+++ app/src/main/java/com/feeltheboard/localflav/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.feeltheboard.eatas
+package com.feeltheboard.localflav
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -6,41 +6,21 @@ import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import com.feeltheboard.eatas.ui.theme.EatasTheme
+import com.feeltheboard.localflav.ui.theme.LocalFlavTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
-            EatasTheme {
-                // A surface container using the 'background' color from the theme
+            LocalFlavTheme {
                 Surface(
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    Greeting("Android")
+                    // call main function
                 }
             }
         }
-    }
-}
-
-@Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    EatasTheme {
-        Greeting("Android")
     }
 }
