+++ app/src/main/java/com/feeltheboard/localflav/components/DishDescriptionCard.kt
@@ -10,8 +10,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import com.feeltheboard.eatas.R
-import com.feeltheboard.localflav.ui.theme.EatasTheme
+import com.feeltheboard.localflav.R
+import com.feeltheboard.localflav.ui.theme.LocalFlavTheme
 
 @Composable
 fun DishDescriptionCard(
@@ -32,7 +32,7 @@ fun DishDescriptionCard(
 @Preview
 @Composable
 fun DishDescriptionPreview() {
-    EatasTheme {
+    LocalFlavTheme {
         DishDescriptionCard(R.string.description_1)
     }
 }
