+++ app/src/main/java/com/feeltheboard/localflav/components/DishImageCard.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import com.feeltheboard.eatas.R
+import com.feeltheboard.localflav.R
 
 @Composable
 fun DishImageCard(
