+++ app/src/main/java/com/feeltheboard/localflav/ui/theme/Theme.kt
@@ -82,7 +82,7 @@ private val DarkColors = darkColorScheme(
 )
 
 @Composable
-fun EatasTheme(
+fun LocalFlavTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     dynamicColor: Boolean = false,
     content: @Composable () -> Unit
